+++ src/util.rs
@@ -31,9 +31,12 @@ pub fn format_ms_timestamp(timestamp_ms: f64) -> String {
 pub fn format_label_ns(plot_name: &str, val: &PlotPoint) -> String {
     let time_ns = val.x;
     let time_s = time_ns / NANOS_PER_SEC as f64;
-    let remainder_ns = time_s.fract() * NANOS_PER_SEC as f64;
-    let dt = DateTime::from_timestamp(time_s as i64, remainder_ns as u32)
-        .unwrap_or_else(|| panic!("Timestamp value out of range: {time_ns}"));
+    let ns_remainder = time_s.fract() * NANOS_PER_SEC as f64;
+    let Some(dt) = DateTime::from_timestamp(time_s as i64, ns_remainder as u32) else {
+        // Will happen if the user zooms out where the X-axis is extended >100 years
+        log::warn!("Timestamp value out of range: {time_s}");
+        return "out of range".to_owned();
+    };
     format!(
         "{plot_name}\ny: {y:.4}\n{h:02}:{m:02}:{s:02}.{subsec_ms:03}",
         y = val.y,