+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.14.2]
+
+### Fixed
+
+- Fix [#250](https://github.com/luftkode/plotinator3000/issues/250) crash when the calculated distance between x-axis grid marks is less than machine epsilon
+- Fix crash on timestamp out of range when X-axis stretches >100 years
 
 ### Dependencies